Removes non-JS jokes
First, I would like to say that this is a terrible way to get Hacktoberfest credits for anyone that is submitting these jokes to this repository (please feel free to mark this PR as `invalid` so I don't get credit for this one either). Second, if the intent of this repository is to collect jokes pertaining to JavaScript, there should be some criteria that the jokes are ACTUALLY relevant to JavaScript. This PR removes all of the jokes that had nothing to do with JavaScript.